+++ sophomore/linear algebra/src/utils/matiter.cpp
@@ -1,7 +1,7 @@
 #include "../include/linalg.h"
 
 // void matIter(Mat *m, void (*f)(Mat*))
-void matiter(Mat *m, void (*f)(int *))
+void matiter(Mat *m, void (*f)(float *))
 {
     for (int i = 0; i < m->row; i++)
         for (int j = 0; j < m->col; j++)
@@ -10,7 +10,7 @@ void matiter(Mat *m, void (*f)(int *))
         }
 }
 
-void matiterij(Mat *m, void (*f)(int *, int i, int j))
+void matiterij(Mat *m, void (*f)(float *, int i, int j))
 {
     for (int i = 0; i < m->row; i++)
         for (int j = 0; j < m->col; j++)
